requirements.txt
decorator==4.4.1
Django==3.0.3
django-ranged-response==0.2.0
django-simple-captcha==0.5.12
et-xmlfile==1.0.1
idna==2.8
jdcal==1.4.1
jedi==0.15.1
mysqlclient==1.4.6
parso==0.5.1
pbr==5.3.0
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.0.0
prompt-toolkit==2.0.10
ptyprocess==0.6.0
Pygments==2.4.2
PyMySQL==0.9.3
pytz==2019.3
requests==2.22.0
sip==4.19.8
six==1.12.0
sqlparse==0.3.0
stevedore==1.30.1
traitlets==4.3.3
django-glrm==1.1.3
